chore(#75): VM base
