+++ app/src/main/kotlin/com/jjdev/equi/Investment.kt
@@ -1,0 +1,10 @@
+package com.jjdev.equi
+
+data class Investment(
+    val ticker: String,
+    val weight: Double,
+    val currentValue: Double,
+    var investedAmount: Double? = null,
+    var targetValue: Double? = null,
+)
+
